requirements.txt
autoawq>=0.2.7
transformers>=4.51.0
torch>=2.4.0
accelerate>=0.30.0
datasets>=2.19.0
jupyter>=1.0.0
ipywidgets>=8.0.0
huggingface_hub>=0.23.0
